backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
python-decouple>=3.8
requests>=2.31
google-generativeai>=0.3
openai>=1.6
psycopg2-binary>=2.9
gunicorn>=21.2
